Task 05.4 Refactoring : Course Model
The course model is supposed to the basic "model"/template in creating a course, those the bare minimum of information is used to make it.

A courseOffering model is introduced to handle more dynamic occurrences that relates to a specific course, i.e. The academic year the course is being offered and the amount of students enrolled into the course for that specific academic year and semester

1. Removed semester and aNum attributes
2. Introduced courseOffering model, to dynamically handle totalStudentsEnrolled and the academic year of the course.
3. Adjusted init and get_json functions
4. Introduced str and repr