Implemented Degree Type ClashRule
Did some overhauling of previous CLI commands specifically the add-assessment command in the course group. Within there I configured it to use different validation helper functions.

I refactored the validate by degree code to suit these updates, Using an enum to track the clashRule type that's stored within the courseAssessment Model.

Introducted the prettyTable module, to simulate a calendar and written up code for the calendar to be generated after a user inputs the courseCode of the course they want to schedule, this allow the system to pinpoint what degree the course is from to then look for all the assessments for other courses within that programme that are already scheduled  to mark those days with an [X].

Initialize.py have some testing code, I used because creating everything from scratch was driving me insane just to see an error lol.

The unit and integration tests have not been updated with these changes for those test which uses the affected models and controllers