Implemented AssessmentType ClashRule
Introducted the assessmentType ClashRule, in order to do this strict measures had to be put in place for the assessment types into entered in the system. Thus, just like was done for the clashRules, and enum was used to store the types of assessments accepted.

Other functioins/CLI commands that relied on the old way (just typing any string) was refactored to adhere to this new standard.

With all the rules applied, and demonstratable via CLI the only thing left is heavy testing to ensure all bases are covered and to add more commands to suit fostering it a bit more like a "view course schedule" maybe a "re-schedule course" etc to make more solid, will need some time away from code to think about it