API Views Implementation: Refactoring
When all the different API Views merged into one branch and some incomplete or incorrect both in views code and Postman scripts, I've fixed all of them (both implementation and Postman scripts) to achieve a fully running and passing Postman Collection.

Individual commits was not made for this due to the continuous trail and error and painstaking long testing of each and every failing test, plus I forgot because I was tired and overwhelmed.